+++ pkg/envoy/xdsutil_test.go
@@ -14,6 +14,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
+	"github.com/openservicemesh/osm/pkg/envoy/secrets"
 	"github.com/openservicemesh/osm/pkg/identity"
 	"github.com/openservicemesh/osm/pkg/tests"
 )
@@ -104,75 +105,6 @@ var _ = Describe("Test Envoy tools", func() {
 		})
 	})
 
-	Context("Test UnmarshalSDSCert()", func() {
-		It("Interface marshals and unmarshals preserving the exact same data", func() {
-			InitialObj := SDSCert{
-				CertType: ServiceCertType,
-				Name:     "test-namespace/test-service",
-			}
-
-			// Marshal/stringify it
-			marshaledStr := InitialObj.String()
-
-			// Unmarshal it back from the string
-			finalObj, _ := UnmarshalSDSCert(marshaledStr)
-
-			// First and final object must be equal
-			Expect(*finalObj).To(Equal(InitialObj))
-		})
-
-		It("returns service cert", func() {
-			actual, err := UnmarshalSDSCert("service-cert:namespace-test/blahBlahBlahCert")
-			Expect(err).ToNot(HaveOccurred())
-			Expect(actual.CertType).To(Equal(ServiceCertType))
-			Expect(actual.Name).To(Equal("namespace-test/blahBlahBlahCert"))
-		})
-		It("returns root cert for mTLS", func() {
-			actual, err := UnmarshalSDSCert("root-cert-for-mtls-outbound:namespace-test/blahBlahBlahCert")
-			Expect(err).ToNot(HaveOccurred())
-			Expect(actual.CertType).To(Equal(RootCertTypeForMTLSOutbound))
-			Expect(actual.Name).To(Equal("namespace-test/blahBlahBlahCert"))
-
-		})
-
-		It("returns root cert for non-mTLS", func() {
-			actual, err := UnmarshalSDSCert("root-cert-https:namespace-test/blahBlahBlahCert")
-			Expect(err).ToNot(HaveOccurred())
-			Expect(actual.CertType).To(Equal(RootCertTypeForHTTPS))
-			Expect(actual.Name).To(Equal("namespace-test/blahBlahBlahCert"))
-		})
-
-		It("returns an error (invalid formatting)", func() {
-			_, err := UnmarshalSDSCert("blahBlahBlahCert")
-			Expect(err).To(HaveOccurred())
-		})
-
-		It("returns an error (invalid formatting)", func() {
-			_, err := UnmarshalSDSCert("blahBlahBlahCert:moreblabla/amazingservice:bla")
-			Expect(err).To(HaveOccurred())
-		})
-
-		It("returns an error (missing cert type)", func() {
-			_, err := UnmarshalSDSCert("blahBlahBlahCert/service")
-			Expect(err).To(HaveOccurred())
-		})
-
-		It("returns an error (invalid serv type)", func() {
-			_, err := UnmarshalSDSCert("revoked-cert:blah/BlahBlahCert")
-			Expect(err).To(HaveOccurred())
-		})
-
-		It("returns an error (invalid mtls cert type)", func() {
-			_, err := UnmarshalSDSCert("oot-cert-for-mtls-diagonalstream:blah/BlahBlahCert")
-			Expect(err).To(HaveOccurred())
-		})
-
-		It("returns an error (empty slice)", func() {
-			_, err := UnmarshalSDSCert(":")
-			Expect(err).To(HaveOccurred())
-		})
-	})
-
 	Context("Test GetDownstreamTLSContext()", func() {
 		It("should return TLS context", func() {
 			svcAccount := identity.K8sServiceAccount{Name: "foo", Namespace: "test"}
@@ -196,9 +128,9 @@ var _ = Describe("Test Envoy tools", func() {
 					}},
 					ValidationContextType: &auth.CommonTlsContext_ValidationContextSdsSecretConfig{
 						ValidationContextSdsSecretConfig: &auth.SdsSecretConfig{
-							Name: SDSCert{
+							Name: secrets.SDSCert{
 								Name:     "test/foo",
-								CertType: RootCertTypeForMTLSInbound,
+								CertType: secrets.RootCertTypeForMTLSInbound,
 							}.String(),
 							SdsConfig: &core.ConfigSource{
 								ConfigSourceSpecifier: &core.ConfigSource_Ads{
@@ -306,13 +238,13 @@ var _ = Describe("Test Envoy tools", func() {
 
 	Context("Test getCommonTLSContext()", func() {
 		It("returns proper auth.CommonTlsContext for outbound mTLS", func() {
-			tlsSDSCert := SDSCert{
+			tlsSDSCert := secrets.SDSCert{
 				Name:     "default/bookbuyer",
-				CertType: ServiceCertType,
+				CertType: secrets.ServiceCertType,
 			}
-			peerValidationSDSCert := SDSCert{
+			peerValidationSDSCert := secrets.SDSCert{
 				Name:     "default/bookstore-v1",
-				CertType: RootCertTypeForMTLSOutbound,
+				CertType: secrets.RootCertTypeForMTLSOutbound,
 			}
 
 			actual := getCommonTLSContext(tlsSDSCert, peerValidationSDSCert)
@@ -336,13 +268,13 @@ var _ = Describe("Test Envoy tools", func() {
 		})
 
 		It("returns proper auth.CommonTlsContext for inbound mTLS", func() {
-			tlsSDSCert := SDSCert{
+			tlsSDSCert := secrets.SDSCert{
 				Name:     "default/bookstore-v1",
-				CertType: ServiceCertType,
+				CertType: secrets.ServiceCertType,
 			}
-			peerValidationSDSCert := SDSCert{
+			peerValidationSDSCert := secrets.SDSCert{
 				Name:     "default/bookstore-v1",
-				CertType: RootCertTypeForMTLSInbound,
+				CertType: secrets.RootCertTypeForMTLSInbound,
 			}
 
 			actual := getCommonTLSContext(tlsSDSCert, peerValidationSDSCert)
@@ -366,13 +298,13 @@ var _ = Describe("Test Envoy tools", func() {
 		})
 
 		It("returns proper auth.CommonTlsContext for non-mTLS (HTTPS)", func() {
-			tlsSDSCert := SDSCert{
+			tlsSDSCert := secrets.SDSCert{
 				Name:     "default/bookstore-v1",
-				CertType: ServiceCertType,
+				CertType: secrets.ServiceCertType,
 			}
-			peerValidationSDSCert := SDSCert{
+			peerValidationSDSCert := secrets.SDSCert{
 				Name:     "default/bookstore-v1",
-				CertType: RootCertTypeForHTTPS,
+				CertType: secrets.RootCertTypeForHTTPS,
 			}
 
 			actual := getCommonTLSContext(tlsSDSCert, peerValidationSDSCert)
