+++ pkg/service/errors.go
@@ -5,5 +5,7 @@ import (
 )
 
 var (
-	errInvalidMeshServiceFormat = errors.New("invalid namespaced service string format")
+	// ErrInvalidMeshServiceFormat indicates the format of the information used to populate
+	// the MeshService is invalid
+	ErrInvalidMeshServiceFormat = errors.New("invalid namespaced service string format")
 )
