+++ pkg/service/service.go
@@ -14,23 +14,3 @@ func (ms MeshService) Equals(service MeshService) bool {
 func (ms MeshService) ServerName() string {
 	return strings.Join([]string{ms.Name, ms.Namespace, "svc", "cluster", "local"}, ".")
 }
-
-// UnmarshalMeshService unmarshals a NamespaceService type from a string
-func UnmarshalMeshService(str string) (*MeshService, error) {
-	slices := strings.Split(str, namespaceNameSeparator)
-	if len(slices) != 2 {
-		return nil, errInvalidMeshServiceFormat
-	}
-
-	// Make sure the slices are not empty. Split might actually leave empty slices.
-	for _, sep := range slices {
-		if len(sep) == 0 {
-			return nil, errInvalidMeshServiceFormat
-		}
-	}
-
-	return &MeshService{
-		Namespace: slices[0],
-		Name:      slices[1],
-	}, nil
-}
